fix(externalStorage): Optimize the used task pop-up window

+++ apps/daas/src/views/external-storage/List.vue
@@ -126,6 +126,25 @@
         </li>
       </ul>
     </Drawer>
+    <el-dialog :visible.sync="showUsingTaskDialog" title="提示">
+      <div>该外存已被个任务调用，请删除或者修改配置后重试</div>
+      <el-table height="250px" :data="usingTasks">
+        <el-table-column min-width="240" :label="$t('public_task_name')" :show-overflow-tooltip="true">
+          <template #default="{ row }">
+            <span class="dataflow-name link-primary flex">
+              <ElLink
+                role="ellipsis"
+                type="primary"
+                class="justify-content-start ellipsis block"
+                :class="['name']"
+                @click.stop="handleClickName(row)"
+                >{{ row.name }}</ElLink
+              >
+            </span>
+          </template>
+        </el-table-column>
+      </el-table>
+    </el-dialog>
   </section>
 </template>
 <script>
@@ -172,7 +191,9 @@ export default {
       isShowDetails: false,
       details: '',
       info: [],
-      labelWidth: '120px'
+      labelWidth: '120px',
+      showUsingTaskDialog: false,
+      usingTasks: []
     }
   },
   computed: {
@@ -315,13 +336,19 @@ export default {
       })
     },
     async remove(row) {
+      //先去请求是否外存已被使用了
+      this.usingTasks = await await externalStorageApi.usingTask(row.id)
       const flag = await this.$confirm(i18n.t('daas_external_storage_list_querenshanchuwai'), '', {
         type: 'warning',
         showClose: false
       })
       if (flag) {
-        await externalStorageApi.delete(row.id)
-        this.table.fetch()
+        if (this.usingTasks) {
+          this.showUsingTaskDialog = true
+        } else {
+          await externalStorageApi.delete(row.id)
+          this.table.fetch()
+        }
       }
     },
     checkDetails(row) {
@@ -339,6 +366,27 @@ export default {
         { label: this.$t('daas_external_storage_list_sheweimoren'), value: row.defaultStorage, icon: 'record' }
       ]
       this.isShowDetails = true
+    },
+    /**
+     * 点击名称调整
+     * @param row
+     */
+    handleClickName(item) {
+      if (item?.syncType === 'migrate') {
+        this.$router.push({
+          name: 'migrateList',
+          query: {
+            keyword: item.name
+          }
+        })
+      } else {
+        this.$router.push({
+          name: 'dataflowList',
+          query: {
+            keyword: item.name
+          }
+        })
+      }
     }
   }
 }
